Revert "you fucking fucked the solutions i reverted to my own push fucking check for errors befor pushing do you even program what have you added in last few days"
This reverts commit d61deb4354af46a71fa1fdaa0112e7f6f39f124f.

+++ Physics/Object.cpp
@@ -166,7 +166,7 @@ namespace Physics
 		//====================================================//
 	}
 
-	void PhysicsObject::SetVelocity(long double x, long double y, long double z)
+	void PhysicsObject::SetVelocity(float x, float y, float z)
 	{
 		m_Velocity.build(x, y, z);
 	}
@@ -176,7 +176,7 @@ namespace Physics
 		m_Velocity = vec;
 	}
 
-	void PhysicsObject::AddVelocity(long double x, long double y, long double z)
+	void PhysicsObject::AddVelocity(float x, float y, float z)
 	{
 		m_Velocity.x += x;
 		m_Velocity.y += y;
@@ -190,7 +190,7 @@ namespace Physics
 		m_Velocity.z += vel.z;
 	}
 
-	void PhysicsObject::SetPosition(long double x, long double y, long double z)
+	void PhysicsObject::SetPosition(float x, float y, float z)
 	{
 		m_Position.build(x, y, z);
 	}
