+++ Physics/Object.h
@@ -27,8 +27,8 @@ namespace Physics
 		void DLLEXPORT SetForce(const vector& Force);
 
 		void DLLEXPORT SetVelocity(const vector& vel);
-		void DLLEXPORT SetVelocity(long double x, long double y, long double z);
-		void DLLEXPORT AddVelocity(long double x, long double y, long double z);
+		void DLLEXPORT SetVelocity(float x, float y, float z);
+		void DLLEXPORT AddVelocity(float x, float y, float z);
 		void DLLEXPORT AddVelocity(const vector& vel);
 
 		void DLLEXPORT SetGravity(long double x, long double y, long double z);
@@ -43,7 +43,7 @@ namespace Physics
 		//This PointerAddress must be ACTUAL Object's position
 		//So, if we update position with our physics code, then the actual position of the object
 		//will be updated too
-		void DLLEXPORT SetPosition(long double x, long double y, long double z);
+		void DLLEXPORT SetPosition(float x, float y, float z);
 		point DLLEXPORT GetPosition();
 
 	public:
